refactor(core): split turn cancellation persistence
